src/requirements.txt
discord==1.0.1
feedparser==6.0.2
numpy==1.20.3
python-dotenv==0.17.1
python_dateutil==2.8.1
pytz==2021.1
skimage==0.0
